Pivot And Bomb Created

+++ EscapeTheCave/MiningPoint.h
@@ -7,7 +7,7 @@
 
 class MiningPoint : public Object {
     private:
-        int distanceToPlayer;
+        float distanceToPlayer;
         Player * player = nullptr;
         Stone * collidingStone = nullptr;
